+++ packages/components/src/components/tooltip/tooltip.custom.styles.ts
@@ -1,0 +1,8 @@
+import { css } from 'lit';
+
+export default css`
+  /* Write custom CSS here */
+  .tooltip__body {
+    box-shadow: var(--syn-shadow-large);
+  }
+`;