+++ packages/components/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@synergy-design-system/components-v1.21.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/1.20.2...components/1.21.0) (2024-04-26)
+
+
+### Features
+
+* ✨ tooltip ([#451](https://github.com/synergy-design-system/synergy-design-system/issues/451)) ([b6f3958](https://github.com/synergy-design-system/synergy-design-system/commit/b6f395846b00598273d0287daa98a64f82b75699))
+
 # [@synergy-design-system/components-v1.20.2](https://github.com/synergy-design-system/synergy-design-system/compare/components/1.20.1...components/1.20.2) (2024-04-26)